+++ watch.sh
@@ -1,0 +1,2 @@
+# https://lib.rs/crates/cargo-watch
+cargo watch -q -c -x 'expand wf_field::definitions'
