+++ wf_common/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "wf_common"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+regex = "1.5.5"
