+++ wf_common/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod common;
+pub mod constants;
+
+#[cfg(test)]
+mod test;
